backend/requirements.txt
fastapi>=0.115.6
uvicorn[standard]>=0.32.1
sqlalchemy>=2.0.36
pydantic>=2.10.3
pydantic-settings>=2.6.1
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt<4.1
python-multipart>=0.0.17
pandas>=2.2.3
openpyxl>=3.1.5
email-validator>=2.2.0
